Find First and Last Position of Element in Sorted Array
In this question a sorted Array is given. We have to find first and last 
position of occurrence of an element.
We can solve this by using linear search but the time complexity is O(n).We can optimize this solution because the given array is sorted
we can use binary search which have time complexity O(logn ).